+++ lib/frp/src/debug.rs
@@ -1,3 +1,4 @@
+//! Root module for debugging utilities.
 
 pub mod graphviz;
