+++ lib/frp/src/io.rs
@@ -1,0 +1,5 @@
+//! Root module for Input / Output FRP bindings
+
+pub mod mouse;
+
+pub use mouse::*;
